Reorder if statements. Check probable scenarios first
The if statements to check whether a record is valid or not should
start by checking the most probable scenario first. Edge cases
should be tested further down the test stack.

When you consider the csv files that will be uploaded, one can safely
say that it will have more valid records than invalid ones. In fact
we might not get invalid records at all. If this is the case then
it makes sense to check for valid records first and then moving on
than to be checking for other scenarios