Add ability to check for extra contingent fields
In some situations, some endpoints require extra fields in addition to their contingent fields. These extra fields are not always required. They are required in some situations only. When those situations arise. We set the extra field in the FayaSMS struct to true, this will cause the checks to include the extra contingent fields when checking